+++ Assets/Neoxider/Scripts/Animations/ColorAnimator.cs
@@ -148,6 +148,7 @@ namespace Neo.Animations
         }
 
         /// <summary>Pauses the animation.</summary>
+        [Button(PlayModeOnly = true)]
         public void Pause()
         {
             if (IsPlaying)
@@ -158,6 +159,7 @@ namespace Neo.Animations
         }
 
         /// <summary>Resumes the animation from pause.</summary>
+        [Button(PlayModeOnly = true)]
         public void Resume()
         {
             if (IsPaused)