+++ Assets/Neoxider/Scripts/Animations/FloatAnimator.cs
@@ -156,6 +156,7 @@ namespace Neo.Animations
         }
 
         /// <summary>Pauses the animation.</summary>
+        [Button(PlayModeOnly = true)]
         public void Pause()
         {
             if (IsPlaying)
@@ -166,6 +167,7 @@ namespace Neo.Animations
         }
 
         /// <summary>Resumes the animation from pause.</summary>
+        [Button(PlayModeOnly = true)]
         public void Resume()
         {
             if (IsPaused)