+++ Assets/Neoxider/Scripts/Tools/Components/TypewriterEffectComponent.cs
@@ -203,6 +203,7 @@ namespace Neo.Tools
         /// <summary>
         ///     Clears text.
         /// </summary>
+        [Button(PlayModeOnly = true)]
         public void Clear()
         {
             Stop();